optimized code for verification of form schemas, created types for all tables

+++ src/app/api/confirmation/[token]/route.tsx
@@ -1,7 +1,8 @@
 import { NextResponse } from 'next/server';
 import getConnection from '@/utility/dbHandler';
-import  {env} from '../../../../utility/EnvironmentValidatior';
+import { env } from '../../../../utility/EnvironmentValidatior';
 import * as jose from 'jose';
+import { User } from '@/types/TableModels';
 export async function GET(request: Request, { params }: { params: { token: string } }) {
     try {
         const emailToken = params.token;
@@ -10,8 +11,8 @@ export async function GET(request: Request, { params }: { params: { token: strin
         );
         const user = await jose.jwtVerify(emailToken, secret);
         const database = getConnection();
-        if (database) {
-            await database('users').where({ id: user.payload.id }).update({
+        if (database && typeof user.payload.id === "string") {
+            await database<User>('users').where({ id: parseInt(user.payload.id) }).update({
                 confirmation: true
             })
             await database.destroy();
